+++ examples/dialog.js
@@ -26,14 +26,15 @@ const textbox = add([
 	rect(width() - 200, 120, { radius: 32 }),
 	anchor("center"),
 	pos(center().x, height() - 100),
-	outline(2),
+	outline(4),
 ])
 
 // Text
 const txt = add([
-	text("", { size: 32, width: width() - 230 }),
+	text("", { size: 32, width: width() - 230, align: "center" }),
 	pos(textbox.pos),
 	anchor("center"),
+	color(0, 0, 0),
 ])
 
 // Character avatar